chore: workaround playwright tsx transpilation
